+++ berks-cookbooks/runit/attributes/default.rb
@@ -2,7 +2,7 @@
 # Cookbook Name:: runit
 # Attribute File:: sv_bin
 #
-# Copyright 2008-2009, Chef Software, Inc.
+# Copyright 2008-2016, Chef Software, Inc.
 #
 # Licensed under the Apache License, Version 2.0 (the 'License');
 # you may not use this file except in compliance with the License.
@@ -47,16 +47,4 @@ when 'rhel', 'fedora'
   default['runit']['start'] = '/etc/init.d/runit-start start'
   default['runit']['stop'] = '/etc/init.d/runit-start stop'
   default['runit']['reload'] = '/etc/init.d/runit-start reload'
-
-when 'gentoo'
-  default['runit']['sv_bin'] = '/usr/bin/sv'
-  default['runit']['chpst_bin'] = '/usr/bin/chpst'
-  default['runit']['service_dir'] = '/var/service'
-  default['runit']['sv_dir'] = '/etc/sv'
-  default['runit']['lsb_init_dir'] = '/etc/init.d'
-  default['runit']['executable'] = '/sbin/runit'
-  default['runit']['start'] = '/etc/init.d/runit-start start'
-  default['runit']['stop'] = '/etc/init.d/runit-start stop'
-  default['runit']['reload'] = '/etc/init.d/runit-start reload'
-
 end