+++ berks-cookbooks/seven_zip/resources/archive.rb
@@ -26,3 +26,4 @@ attribute :path, kind_of: String, name_attribute: true
 attribute :source, kind_of: String
 attribute :overwrite, kind_of: [TrueClass, FalseClass], default: false
 attribute :checksum, kind_of: String
+attribute :timeout, kind_of: Integer