+++ berks-cookbooks/windows/attributes/default.rb
@@ -3,7 +3,7 @@
 # Cookbook Name:: windows
 # Attribute:: default
 #
-# Copyright 2011-2015, Chef Software, Inc
+# Copyright 2011-2016, Chef Software, Inc
 #
 # Licensed under the Apache License, Version 2.0 (the "License");
 # you may not use this file except in compliance with the License.
@@ -18,7 +18,4 @@
 # limitations under the License.
 #
 
-default['windows']['allow_pending_reboots'] = true
-default['windows']['allow_reboot_on_failure'] = false
 default['windows']['rubyzipversion'] = nil
-default['windows']['reboot_timeout'] = 60