+++ berks-cookbooks/ark/CHANGELOG.md
@@ -2,9 +2,17 @@
 
 This file is used to list changes made in each version of the ark cookbook.
 
-## [1.2](https://github.com/chef-cookbooks/ark/tree/1.2) (2016-07-03)
+## 2.0.0 (2016-09-15)
 
-[Full Changelog](https://github.com/chef-cookbooks/ark/compare/v1.1.0...1.2)
+- Add CentOS 7.2, Fedora 23, and Suse specs
+- Add centos 5, debian, and opensuse travis testing
+- Add a contributing doc
+- Fix cookstyle warnings
+- Require Chef 12.1+
+
+## [v1.2.0](https://github.com/chef-cookbooks/ark/tree/v1.2.0) (2016-07-03)
+
+[Full Changelog](https://github.com/chef-cookbooks/ark/compare/v1.1.0...v1.2.0)
 
 - Create seven_zip unpack command when strip_components is 0 [#155](https://github.com/chef-cookbooks/ark/pull/155) ([terkill](https://github.com/terkill))
 - Get 7zip path from the windows registry. [#153](https://github.com/chef-cookbooks/ark/pull/153) ([buri17](https://github.com/buri17))