+++ berks-cookbooks/yum-epel/CHANGELOG.md
@@ -2,6 +2,19 @@
 
 This file is used to list changes made in each version of the yum-epel cookbook.
 
+## 1.0.1 (2016-09-11)
+- Fix epel-testing attributes
+
+## 1.0.0 (2016-09-06)
+- Add chef_version metadata
+- Testing updates
+- Remove support for Chef 11
+
+## v0.7.1 (2016-08-19)
+- Remove bats testing
+- Fix attribute settings 
+- Cleanup travis file
+
 ## v0.7.0 (2016-04-27)
 
 - Added support for IBM zlinux platform